added working but random robot pos grid and logger
